Performed a full code audit
Numerous bugs have been fixed, thus, the performance and reliability of the application is better than ever.

+++ MoneroGui/Views/MainWindow/StatusBarViewModel.cs
@@ -52,8 +52,8 @@ namespace Jojatekok.MoneroGUI.Views.MainWindow
             }
         }
 
-        private byte _connectionCount;
-        public byte ConnectionCount {
+        private ushort _connectionCount;
+        public ushort ConnectionCount {
             get { return _connectionCount; }
 
             set {
